+++ loading-button-android/src/main/java/br/com/simplepass/loading_button_lib/customViews/CircularProgressButton.kt
@@ -4,20 +4,18 @@ import android.animation.Animator
 import android.animation.AnimatorListenerAdapter
 import android.animation.AnimatorSet
 import android.content.Context
-import android.content.res.TypedArray
 import android.graphics.Bitmap
 import android.graphics.Canvas
 import android.graphics.drawable.GradientDrawable
 import android.util.AttributeSet
 import androidx.appcompat.widget.AppCompatButton
 import androidx.core.content.ContextCompat
-import br.com.simplepass.loading_button_lib.R
 import br.com.simplepass.loading_button_lib.animatedDrawables.CircularProgressAnimatedDrawable
 import br.com.simplepass.loading_button_lib.animatedDrawables.CircularRevealAnimatedDrawable
 import br.com.simplepass.loading_button_lib.disposeAnimator
-import br.com.simplepass.loading_button_lib.presentation.*
+import br.com.simplepass.loading_button_lib.presentation.ProgressButtonPresenter
 
-class CircularProgressButton : AppCompatButton {
+class CircularProgressButton : AppCompatButton, ProgressButton {
 
     constructor(context: Context) : super(context) {
         init()
@@ -31,16 +29,16 @@ class CircularProgressButton : AppCompatButton {
         init(attrs, defStyleAttr)
     }
 
-    private var paddingProgress = 0F //Todo: should be in View
+    override var paddingProgress = 0F
 
-    private var spinningBarWidth = 10F
-    private var spinningBarColor = ContextCompat.getColor(context, android.R.color.black)
+    override var spinningBarWidth = 10F
+    override var spinningBarColor = ContextCompat.getColor(context, android.R.color.black)
 
-    private var finalCorner = 0F
-    private var initialCorner = 0F
+    override var finalCorner = 0F
+    override var initialCorner = 0F
 
-    internal var doneFillColor: Int = ContextCompat.getColor(context, android.R.color.black)
-    internal lateinit var doneImage: Bitmap
+    override var doneFillColor: Int = ContextCompat.getColor(context, android.R.color.black)
+    override lateinit var doneImage: Bitmap
 
     private val finalWidth: Int by lazy { finalHeight }
     var initialWidth = 0
@@ -48,8 +46,13 @@ class CircularProgressButton : AppCompatButton {
     private val finalHeight: Int by lazy { height }
     private val initialHeight: Int by lazy { height }
 
-    lateinit var drawable: GradientDrawable
+    override val viewHeight: Int = height
+    override val viewWidth: Int = width
+
+    override lateinit var drawable: GradientDrawable
     val initialText: CharSequence = text
+
+    private val presenter = ProgressButtonPresenter(this)
 
     internal val morphAnimator by lazy {
         AnimatorSet().apply {
@@ -75,63 +78,12 @@ class CircularProgressButton : AppCompatButton {
         }
     }
 
-    private val presenter = ProgressButtonPresenter(this)
-
     private val progressAnimatedDrawable: CircularProgressAnimatedDrawable by lazy {
-        CircularProgressAnimatedDrawable(this, spinningBarWidth, spinningBarColor).apply {
-            val offset = (width - height) / 2
-
-            val left = offset + paddingProgress.toInt()
-            val right = width - offset - paddingProgress.toInt()
-            val bottom = height - paddingProgress.toInt()
-            val top = paddingProgress.toInt()
-
-            setBounds(left, top, right, bottom)
-            callback = this@CircularProgressButton
-        }
+        createProgressDrawable()
     }
 
     private val revealAnimatedDrawable: CircularRevealAnimatedDrawable by lazy {
-        CircularRevealAnimatedDrawable(this, doneFillColor, doneImage).apply {
-            setBounds(0, 0, width, height)
-            callback = this@CircularProgressButton
-        }
-    }
-
-    private fun init(attrs: AttributeSet? = null, defStyleAttr: Int = 0) {
-        val typedArray: TypedArray? = attrs?.run {
-            context.obtainStyledAttributes(
-                this,
-                R.styleable.CircularProgressButton,
-                defStyleAttr,
-                0
-            )
-        }
-
-        val typedArrayBg: TypedArray? = attrs?.run {
-            val attrsArray = intArrayOf(android.R.attr.background)
-            context.obtainStyledAttributes(this, attrsArray, defStyleAttr, 0)
-        }
-
-        drawable = parseGradientDrawable(
-            typedArrayBg?.getDrawable(0)
-                ?: ContextCompat.getDrawable(context, R.drawable.shape_default)!!
-        )
-
-        typedArray?.let { tArray ->
-            presenter.run {
-                initialCorner = tArray.getDimension(R.styleable.CircularProgressButton_initialCornerAngle, 0f)
-                finalCorner = tArray.getDimension(R.styleable.CircularProgressButton_finalCornerAngle, 100f)
-
-                spinningBarWidth = tArray.getDimension(R.styleable.CircularProgressButton_spinning_bar_width, 10f)
-                spinningBarColor = tArray.getColor(R.styleable.CircularProgressButton_spinning_bar_color, spinningBarColor)
-
-                paddingProgress = tArray.getDimension(R.styleable.CircularProgressButton_spinning_bar_padding, 0F)
-            }
-        }
-
-        typedArray?.recycle()
-        typedArrayBg?.recycle()
+        createRevealAnimatedDrawable()
     }
 
     internal fun drawProgress(canvas: Canvas) {
